\name{.Scopus_ImportSources_Countries}
\alias{.Scopus_ImportSources_Countries}
\title{/internal/...}
\usage{.Scopus_ImportSources_Countries(conn, verbose)}
\description{/internal/
Imports country list form Scopus_SourceList to Biblio_Countries
Returns Countries converted to IdCountry for direct use in further queries.}
