% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.info.R
\name{as.character.authorinfo}
\alias{as.character.authorinfo}
\title{Coerce an authorinfo object to character string}
\usage{
\method{as.character}{authorinfo}(x, ...)
}
\arguments{
\item{x}{a single object of class \code{authorinfo}.}

\item{...}{unused.}
}
\value{
A character string
}
\description{
Converts an object of class \code{authorinfo} to a character string.
Such an object is returned by e.g. \code{\link{lbsGetInfoAuthors}}.
}
\details{
An \code{authorinfo} object  is a list with the following components:
\itemize{
\item \code{IdAuthor} --- numeric; author's identifier in the table \code{Biblio_Authors},
\item \code{Name} --- character; author's name.
}
}
\seealso{
\code{\link{print.authorinfo}}, \code{\link{lbsSearchAuthors}}, \code{\link{lbsGetInfoAuthors}}
}

