% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.getcitations.R
\name{lbsGetCitations}
\alias{lbsGetCitations}
\title{Fetch authors' citation sequences}
\usage{
lbsGetCitations(
  conn,
  documentTypes = NULL,
  surveyDescription = NULL,
  idAuthors = NULL,
  verbose = TRUE
)
}
\arguments{
\item{conn}{a connection object as produced by \code{\link{lbsConnect}}.}

\item{documentTypes}{character vector or \code{NULL}; specifies document types to restrict to;
a combination of \code{Article}, \code{Article in Press}, \code{Book}, \code{Conference Paper},
\code{Editorial}, \code{Erratum}, \code{Letter}, \code{Note}, \code{Report}, \code{Review},
\code{Short Survey}. \code{NULL} means no restriction.}

\item{surveyDescription}{single character string or \code{NULL}; survey to restrict to or \code{NULL} for no restriction.}

\item{idAuthors}{numeric vector of authors' identifiers for which the sequences are to be created or \code{NULL} for all authors in the database.}

\item{verbose}{logical; \code{TRUE} to inform about the progress of the process.}
}
\value{
A list of non-increasingly ordered numeric vectors is returned. Each element of the list corresponds
to a citation sequence of some author. List \code{names} attribute are
set to authors' names. Moreover, each vector has a set \code{IdAuthor}
attribute, which uniquely identifies the corresponding record in the table \code{Biblio_Authors}.
Citation counts come together with \code{IdDocument}s (vector elements are named).

The list of citation sequences may then be used to calculate
authors' impact using \code{\link{lbsAssess}} (see Examples below).
}
\description{
Creates ordered citation sequences of authors in a Local Bibliometric Storage.
}
\details{
A citation sequence is a numeric vector
consisting of citation counts of all the documents mapped to
selected  authors.
However, the function may take into account only the documents
from a given Survey (using \code{surveyDescription}
parameter) or of chosen types (\code{documentTypes}).
}
\examples{
\dontrun{
conn <- lbsConnect("Bibliometrics.db");
## ...
citseq <- lbsGetCitations(conn,
	surveyDescription="Scientometrics", documentTypes="Article",
	idAuthors=c(39264,39265,39266));
print(citseq);
## $`Liu X.`                                # Author name
## 40116 34128 39122 29672 32343 32775      # IdDocument
##    11     4     1     0     0     0      # Citation count
## attr(,"IdAuthor")
## [1] 39264                                # IdAuthor
##
## $`Xu Y.`
## 38680 38605 40035 40030 40124 39829 39745 29672
##    30    14     8     6     6     5     3     0
## attr(,"IdAuthor")
## [1] 39265
##
## $`Wang Y.`
## 29992 29672 29777 32906 33858 33864 34704
##     1     0     0     0     0     0     0
## attr(,"IdAuthor")
## [1] 39266
print(lbsAssess(citseq,
   f=list(length, sum, index.h, index.g, function(x) index.rp(x,1),
       function(x) sqrt(prod(index.lp(x,1))),
       function(x) sqrt(prod(index.lp(x,Inf)))),
   captions=c("length", "sum", "index.h", "index.g", "index.w",
   "index.lp1", "index.lpInf")));
##      Name length sum index.h index.g index.w index.lp1 index.lpInf
## 3   Xu Y.      8  72       5       8       7  8.573214    5.477226
## 2 Wang Y.      7   1       1       1       1  1.000000    1.000000
## 1  Liu X.      6  16       2       4       3  4.157609    3.316625
## ...
dbDisconnect(conn);}
}
\seealso{
\code{\link{lbsConnect}},
\code{\link{lbsAssess}}
}
