\name{print.authorinfo}
\alias{print.authorinfo}
\title{Print an authorinfo object}
\usage{
  print.authorinfo(x, ...)
}
\arguments{
  \item{x}{an object of type \code{authorinfo}.}

  \item{...}{unused.}
}
\description{
  Prints out an object of type \code{authorinfo}. Such an
  object is returned by e.g.
  \code{\link{lbsGetInfoAuthors}}.
}
\details{
  For more information see man page for
  \code{\link{as.character.authorinfo}}.
}
\seealso{
  \code{\link{as.character.authorinfo}},
  \code{\link{lbsSearchAuthors}},
  \code{\link{lbsGetInfoAuthors}}
}

