\name{print.docinfo}
\alias{print.docinfo}
\title{Print a docinfo object}
\usage{
  print.docinfo(x, ...)
}
\arguments{
  \item{x}{an object of type \code{docinfo}.}

  \item{...}{unused.}
}
\description{
  Prints out an object of type \code{docinfo}. Such an
  object is returned by e.g.
  \code{\link{lbsGetInfoDocuments}}.
}
\details{
  For more information see man page for
  \code{\link{as.character.docinfo}}.
}
\seealso{
  \code{\link{as.character.docinfo}},
  \code{\link{lbsSearchDocuments}},
  \code{\link{lbsGetInfoDocuments}}
}

