% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITAN-package.R
\docType{package}
\name{CITAN-package}
\alias{CITAN}
\alias{CITAN-package}
\title{\pkg{CITAN} is a library of functions useful in
quantitative research in the field of scientometrics.}
\description{
The package is deprecated, see \pkg{agop} instead.
}
\details{
For the complete list of functions, call \code{library(help="CITAN")}.
}
\references{
Gagolewski M. (2011). Bibliometric Impact Assessment with R and the CITAN Package,
   Journal of Informetrics 5(4), 678-692.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gagolews/CITAN}
  \item Report bugs at \url{https://github.com/gagolews/CITAN/issues}
}

}
\author{
\strong{Maintainer}: Marek Gagolewski \email{marek@gagolewski.com} (\href{https://orcid.org/0000-0003-0637-6028}{ORCID}) [copyright holder]

}
\keyword{internal}
