% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.info.R
\name{print.authorinfo}
\alias{print.authorinfo}
\title{Print an authorinfo object}
\usage{
\method{print}{authorinfo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{authorinfo}.}

\item{...}{unused.}
}
\description{
Prints out an object of class \code{authorinfo}. Such an object is returned by e.g. \code{\link{lbsGetInfoAuthors}}.
}
\details{
For more information see man page for \code{\link{as.character.authorinfo}}.
}
\seealso{
\code{\link{as.character.authorinfo}}, \code{\link{lbsSearchAuthors}}, \code{\link{lbsGetInfoAuthors}}
}

