% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio.info.R
\name{print.docinfo}
\alias{print.docinfo}
\title{Print a docinfo object}
\usage{
\method{print}{docinfo}(x, ...)
}
\arguments{
\item{x}{an object of class \code{docinfo}.}

\item{...}{unused.}
}
\description{
Prints out an object of class \code{docinfo}. Such an object is returned by e.g. \code{\link{lbsGetInfoDocuments}}.
}
\details{
For more information see man page for \code{\link{as.character.docinfo}}.
}
\seealso{
\code{\link{as.character.docinfo}}, \code{\link{lbsSearchDocuments}}, \code{\link{lbsGetInfoDocuments}}
}

