% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus.data.R
\docType{data}
\name{Scopus_ASJC}
\alias{Scopus_ASJC}
\title{Scopus ASJC (All Science. Journals Classification) classification codes}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
Scopus_ASJC
}
\description{
List of Elsevier's \emph{SciVerse Scopus} ASJC (All Science. Journals Classification)
source classification codes.
}
\details{
Last update: October 2011. The data file is based on the official and publicly available
(no permission needed as stated by Elsevier) Scopus list of covered titles.

It consists of 334 ASJC 4-digit integer codes (column \code{ASJC})
together with their group identifiers (column \code{ASJC_Parent})
and descriptions (column \code{Description}).

ASJC codes are used to classify Scopus sources (see \code{\link{Scopus_SourceList}}).
}
\seealso{
\code{\link{Scopus_SourceList}}, \code{\link{Scopus_ReadCSV}}, \code{\link{Scopus_ImportSources}}
}
\keyword{datasets}
