% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Scopus_ASJC}
\alias{Scopus_ASJC}
\title{Scopus ASJC (All Science. Journals Classification) classification codes}
\format{\preformatted{ NULL
}}
\usage{
Scopus_ASJC
}
\description{
List of Elsevier's \emph{SciVerse Scopus} ASJC (All Science. Journals Classification)
source classification codes.
}
\details{
Last update: October 2011. The data file is based on the official and publicly available
(no permission needed as stated by Elsevier) Scopus list of covered titles,
see \url{http://www.info.sciverse.com/documents/files/scopus-training/resourcelibrary/xls/title_list.xls}.

It consists of 334 ASJC 4-digit integer codes (column \code{ASJC})
together with their group identifiers (column \code{ASJC_Parent})
and descriptions (column \code{Description}).

ASJC codes are used to classify Scopus sources (see \code{\link{Scopus_SourceList}}).
}
\references{
\url{http://www.info.sciverse.com/scopus/scopus-in-detail/facts/}
}
\seealso{
\code{\link{Scopus_SourceList}}, \code{\link{Scopus_ReadCSV}}, \code{\link{Scopus_ImportSources}}
}
\keyword{ASJC,}
\keyword{Scopus,}
\keyword{journal}

