\name{Scopus_ASJC}
\alias{Scopus_ASJC}
\title{Scopus ASJC (All Science. Journals Classification) classification codes}
\description{List of Elsevier's \emph{SciVerse Scopus} ASJC (All Science. Journals Classification)
source classification codes.}
\details{Last update: March 2011. The data file is based on the official and publicly available
(no permission needed-information from Elsevier) \emph{SciVerse Scopus} list of covered titles,
see \url{http://www.info.sciverse.com/documents/files/scopus-training/resourcelibrary/xls/title_list.xls}.

It consists of 334 ASJC 4-digit integer codes (column \code{ASJC})
together with their group identifiers (column \code{ASJC_Group})
and descriptions (column \code{Description}).

It is used to classify journals and other sources (see \code{\link{Scopus_SourceList}}).}
\docType{data}
\seealso{\code{\link{Scopus_SourceList}}, \code{\link{Scopus_ReadCSV}}, \code{\link{Scopus_ImportSources}}}
\author{Marek Gagolewski \email{gagolews@ibspan.waw.pl}}
\references{\url{http://www.info.sciverse.com/scopus/scopus-in-detail/facts/}}
\keyword{Scopus,}
\keyword{ASJC,}
\keyword{journal}
