\name{sqlNumericOrNULL}
\alias{sqlNumericOrNULL}
\title{Get nullable numeric value for use in an SQL query}
\usage{sqlNumericOrNULL(value)}
\description{Converts a single numeric variable to character string or returns \code{"NULL"}.}
\value{The function returns a character string containing a
textual representation of a given numeric variable or \code{"NULL"}.}
\seealso{\code{\link{sqlSwitchOrNULL}}, \code{\link{sqlStringOrNULL}}}
\arguments{\item{value}{value to be processed.}}
